feat(Switch): Replace deprecated implementation
