+++ packages/react-component-library/package.json
@@ -160,7 +160,6 @@
     "@defencedigital/fonts": "^2.80.2",
     "@defencedigital/icon-library": "^2.80.2",
     "@popperjs/core": "^2.9.2",
-    "@types/react-responsive": "^8.0.2",
     "@types/react-toast-notifications": "^2.4.0",
     "autoprefixer": "^10.2.5",
     "classnames": "^2.2.6",
@@ -175,7 +174,6 @@
     "react-indiana-drag-scroll": "^2.0.1",
     "react-merge-refs": "^1.1.0",
     "react-popper": "^2.2.5",
-    "react-responsive": "^8.0.3",
     "react-select": "^4.1.0",
     "react-string-replace": "^0.4.4",
     "react-toast-notifications": "^2.4.0",
