+++ packages/react-component-library/src/components/Switch/Switch.test.tsx
@@ -1,8 +1,11 @@
 import React from 'react'
 import '@testing-library/jest-dom/extend-expect'
 import { render, RenderResult } from '@testing-library/react'
+import { selectors } from '@defencedigital/design-tokens'
 
-import { Switch } from './Switch'
+import { Switch, SwitchOption } from '.'
+
+const { color } = selectors
 
 describe('Switch', () => {
   let wrapper: RenderResult
@@ -19,28 +22,24 @@ describe('Switch', () => {
           label="Switch label"
           name="switch-name"
           onChange={onChangeSpy}
-          options={[
-            { label: 'Day', value: '1' },
-            { label: 'Week', value: '2' },
-            { label: 'Month', value: '3' },
-            { label: 'Year', value: '4' },
-          ]}
           value="3"
-        />
+        >
+          <SwitchOption label="Option 1" value="1" />
+          <SwitchOption label="Option 2" value="2" />
+          <SwitchOption label="Option 3" value="3" />
+        </Switch>
       )
     })
 
     it('should spread arbitrary props', () => {
-      expect(wrapper.getByTestId('switch-wrapper')).toHaveAttribute(
+      expect(wrapper.getByTestId('switch')).toHaveAttribute(
         'data-arbitrary',
         'arbitrary'
       )
     })
 
     it('should drill the custom CSS class', () => {
-      expect(wrapper.getByTestId('switch-wrapper').classList).toContain(
-        'custom-class'
-      )
+      expect(wrapper.getByTestId('switch').classList).toContain('custom-class')
     })
 
     it('should render the label', () => {
@@ -52,33 +51,31 @@ describe('Switch', () => {
       expect(switchInputs[0].getAttribute('name')).toEqual('switch-name')
       expect(switchInputs[1].getAttribute('name')).toEqual('switch-name')
       expect(switchInputs[2].getAttribute('name')).toEqual('switch-name')
-      expect(switchInputs[3].getAttribute('name')).toEqual('switch-name')
     })
 
-    it('should render 4 options', () => {
-      expect(wrapper.getByText('Day')).toBeInTheDocument()
-      expect(wrapper.getByText('Week')).toBeInTheDocument()
-      expect(wrapper.getByText('Month')).toBeInTheDocument()
-      expect(wrapper.getByText('Year')).toBeInTheDocument()
-      expect(wrapper.getAllByTestId('switch-input')).toHaveLength(4)
+    it('should render 3 options', () => {
+      expect(wrapper.getByText('Option 1')).toBeInTheDocument()
+      expect(wrapper.getByText('Option 2')).toBeInTheDocument()
+      expect(wrapper.getByText('Option 3')).toBeInTheDocument()
+      expect(wrapper.getAllByTestId('switch-input')).toHaveLength(3)
     })
 
     it('should set the value', () => {
-      expect(wrapper.getByText('Month')).toHaveStyleRule(
+      expect(wrapper.getByText('Option 3')).toHaveStyleRule(
         'background-color',
-        '#3e5667'
+        color('action', '600')
       )
     })
 
     describe('when the selected option is changed', () => {
       beforeEach(() => {
-        wrapper.getByText('Year').click()
+        wrapper.getByText('Option 1').click()
       })
 
       it('should set the value', () => {
-        expect(wrapper.getByText('Year')).toHaveStyleRule(
+        expect(wrapper.getByText('Option 1')).toHaveStyleRule(
           'background-color',
-          '#3e5667'
+          color('action', '600')
         )
       })
 
@@ -90,30 +87,54 @@ describe('Switch', () => {
 
   describe('when minimal props are specified', () => {
     beforeEach(() => {
-      onChangeSpy = jest.fn()
-
       wrapper = render(
-        <Switch
-          name="switch-name"
-          options={[
-            { label: 'Day', value: '1' },
-            { label: 'Week', value: '2' },
-            { label: 'Month', value: '3' },
-            { label: 'Year', value: '4' },
-          ]}
-        />
+        <Switch name="switch-name">
+          <SwitchOption label="Option 1" value="1" />
+          <SwitchOption label="Option 2" value="2" />
+          <SwitchOption label="Option 3" value="3" />
+        </Switch>
       )
     })
 
     it('should not render the label', () => {
       expect(wrapper.queryAllByTestId('switch-legend')).toHaveLength(0)
     })
+  })
 
-    it('should not set the value', () => {
-      expect(wrapper.getByText('Day').classList).not.toContain('is-active')
-      expect(wrapper.getByText('Week').classList).not.toContain('is-active')
-      expect(wrapper.getByText('Month').classList).not.toContain('is-active')
-      expect(wrapper.getByText('Year').classList).not.toContain('is-active')
+  describe('when the component is disabled', () => {
+    beforeEach(() => {
+      onChangeSpy = jest.fn()
+
+      wrapper = render(
+        <Switch name="switch-name" isDisabled onChange={onChangeSpy}>
+          <SwitchOption label="Option 1" value="1" />
+          <SwitchOption label="Option 2" value="2" />
+          <SwitchOption label="Option 3" value="3" />
+        </Switch>
+      )
+    })
+
+    it('should not invoke the onChange handler when an option is clicked', () => {
+      wrapper.getByText('Option 1').click()
+      expect(onChangeSpy).not.toHaveBeenCalled()
+    })
+  })
+
+  describe('when the component has a falsey child', () => {
+    beforeEach(() => {
+      const showThird = false
+
+      wrapper = render(
+        <Switch name="switch-name">
+          <SwitchOption label="Option 1" value="1" />
+          <SwitchOption label="Option 2" value="2" />
+          {showThird && <SwitchOption label="Option 3" value="3" />}
+        </Switch>
+      )
+    })
+
+    it('should not render the third option', () => {
+      expect(wrapper.queryByText('Option 3')).not.toBeInTheDocument()
     })
   })
 })
