+++ packages/react-component-library/src/components/Switch/index.ts
@@ -1,3 +1,2 @@
-export * from './constants'
-export * from './ResponsiveSwitch'
 export * from './Switch'
+export * from './SwitchOption'
