+++ packages/react-component-library/src/components/Switch/partials/StyledLegend.tsx
@@ -6,5 +6,5 @@ const { color, spacing } = selectors
 export const StyledLegend = styled.legend`
   font-size: inherit;
   color: ${color('neutral', '400')};
-  margin-bottom: ${spacing('4')};
+  margin: 0 0 ${spacing('3')} ${spacing('6')};
 `
