--- packages/react-component-library/src/components/SwitchE/index.ts
@@ -1,2 +1,0 @@
-export * from './SwitchE'
-export * from './SwitchEOption'
