--- packages/react-component-library/src/components/SwitchE/partials/SwitchInput.tsx
@@ -1,5 +1,0 @@
-import styled from 'styled-components'
-
-export const SwitchInput = styled.input`
-  display: none;
-`
