+++ packages/react-component-library/src/index.ts
@@ -42,7 +42,6 @@ export * from './components/RangeSlider'
 export * from './components/SectionDivider'
 export * from './components/Select'
 export * from './components/Switch'
-export * from './components/SwitchE'
 export * from './components/Table'
 export * from './components/TabNav'
 export * from './components/TabSet'
